Fix warning about method not returning on all paths
OgreMaterial::getOgreTechniqueForConfiguration does not return something on
all execution paths. This results in a warning on vs2010. It should either
return NULL or throw an exception if it cannot find the configurationName
and lodIndex. This patch makes it throw an exception as I am not sure if
NULL should be allowed.